refactor(ui-components): reorganized ui lib files
