+++ src/components/ui/button.tsx
@@ -2,7 +2,7 @@ import * as React from 'react';
 import { Slot } from '@radix-ui/react-slot';
 import { type VariantProps } from 'class-variance-authority';
 import { cn } from '@/lib/utils';
-import { buttonVariants } from '@/lib/button-variants';
+import { buttonVariants } from '@/lib/ui/button-variants';
 
 /**
  * Button component props interface
