+++ src/lib/ui/skeleton-variants.ts
